Add HTML5 target option for custom canvas ID
Kha scripts target canvas elements with the "khanvas" ID by default.
Because the canvas ID is hardcoded, it is not possible to somehow specify
different target elements. If that were possible, it would allow for more
HTML5 Kha projects to be present on the same page.

The changes proposed introduce a new HTML5 target option to `khafile.js` -
`canvasId` - which sets the target canvas ID, defaulting to the previously
defined "khanvas" if no ID is provided. The option is used as follows:

`project.targetOptions.html5.canvasId = 'myCustomId';`

The provided option is passed to the compiler as a value to the `canvasId`
flag, which can then be retrieved in the macro context.