Add HTML5 target options for custom script name
The JS scripts generated by Kha always bear the same name - "kha.js". This
name is hardcoded, which means that the generated scripts need to be
renamed manually.

The changes proposed introduce a new HTML5 target option to `khafile.js` -
`scriptName` - which sets the generated script's name (sans extension),
defaulting to "kha" if no value is provided.

This is only implemented for the HTML5 target, and not HTML5 Debug, HTML5
Worker, or Node targets.

The option is used as follows:

`project.targetOptions.html5.scriptName = 'new-project';`

This will generate a script with the name `new-project.js` and use it in
the default template.

The option is also passed to the compiler as a value to the `script_name`
flag, which can be retrieved in the macro context.