--parallelAssetConversion: fix value parsing
The value of the `--parallelAssetConversion` option was treated as a string, which did not work for the `=== -1` comparison that enables the automatic processes amount.

I'm not sure why it worked for positive values before since they were strings as well. Perhaps `Throttle.all()` (called by the assets and shader conversion code) can convert strings, but we shouldn't rely on that.